Add ability to use simple server with the buffered transport (#23)
The nonblocking server with the framed transport is not supported
by the all languages that use remote controller. For example,
nonblocking server is not supported for the Node.js binding and
framed transport is not supported for the Python binding.
Therefore, this PR adds an ability to use simple server with
buffered transport when the `--use-simple-server` flag is
provided.