Add login modules for client security tests
It is not possible to test some of the authentication mechanisms
without writing Java code.

Instead of adding these modules to enterprise-tests JAR, I have
decided to add them here, so that we could use those classes
in all clients&against all v4 and v5 Hazelcast versions.

This PR introduces two separate login modules.

- One that uses custom credentials (called SimpleCredentials), and
its serialization factory.
- One that uses TokenCredentials.

There was no need to introduce a login module to test
UsernamePasswordCredentials as with the default authentication, we
can test the username and password sent by the client.