Add Compact credentials to test custom authentication with Compact
This PR adds a Compact serializable credentials class and associated
login module, so that the custom authentication with Compact
can be tested in non-java clients.

The non-Java clients would need to write a serializer for the
CompactCredentials in their language, and use the fully qualified
class name as the type name (`com.hazelcast.security.CompactCredentials`)
so that the Java can use reflective serializer to deserialize it,
just like other Compact classes put into remote-controller.

To use this login module in non-Java client tests, one has to
use a configuration like below

```xml
<hazelcast xmlns="http://www.hazelcast.com/schema/config" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.hazelcast.com/schema/config http://www.hazelcast.com/schema/config/hazelcast-config-5.2.xsd">
    <security enabled="true">
        <realms>
            <realm name="realm">
                <authentication>
                    <jaas>
                        <login-module class-name="com.hazelcast.security.CompactCredentialsLoginModule" usage="REQUIRED">
                            <properties>
                                <property name="key2">xyz</property>
                                <property name="key1">abc</property>
                                <property name="username">user</property>
                            </properties>
                        </login-module>
                    </jaas>
                </authentication>
            </realm>
        </realms>
        <client-authentication realm="realm"/>
    </security>
</hazelcast>
```